+++ sdks/python/apache_beam/runners/interactive/interactive_runner_test.py
@@ -27,20 +27,11 @@ from __future__ import print_function
 import unittest
 
 import apache_beam as beam
-from apache_beam.runners import runner
 from apache_beam.runners.direct import direct_runner
 from apache_beam.runners.interactive import interactive_beam as ib
 from apache_beam.runners.interactive import interactive_environment as ie
 from apache_beam.runners.interactive import interactive_runner
-from apache_beam.testing.test_stream import TestStream
-from apache_beam.transforms.window import TimestampedValue
 
-# TODO(BEAM-8288): clean up the work-around of nose tests using Python2 without
-# unittest.mock module.
-try:
-  from unittest.mock import patch
-except ImportError:
-  from mock import patch
 
 
 def print_with_message(msg):
@@ -49,17 +40,6 @@ def print_with_message(msg):
     return elem
 
   return printer
-
-def _build_a_test_stream_pipeline():
-  test_stream = (TestStream()
-                 .advance_watermark_to(0)
-                 .add_elements([TimestampedValue('a', 1)])
-                 .advance_processing_time(5)
-                 .advance_watermark_to_infinity())
-  p = beam.Pipeline(runner=interactive_runner.InteractiveRunner())
-  events = p | test_stream  # pylint: disable=possibly-unused-variable
-  ib.watch(locals())
-  return p
 
 
 class InteractiveRunnerTest(unittest.TestCase):
@@ -141,52 +121,6 @@ class InteractiveRunnerTest(unittest.TestCase):
     self.assertTrue(underlying_runner._in_session)
     runner.end_session()
     self.assertFalse(underlying_runner._in_session)
-
-  # A patch is needed until the boundedness check considers a TestStream as
-  # an unbounded source.
-  @patch('apache_beam.runners.interactive.pipeline_instrument.'
-         'PipelineInstrument.has_unbounded_sources', True)
-  def test_background_caching_job_starts_when_none_such_job_exists(self):
-    p = _build_a_test_stream_pipeline()
-    p.run()
-    self.assertIsNotNone(
-        ie.current_env().pipeline_result(p, is_main_job=False))
-
-  @patch('apache_beam.runners.interactive.pipeline_instrument.'
-         'PipelineInstrument.has_unbounded_sources', False)
-  def test_background_caching_job_not_start_for_batch_pipeline(self):
-    p = _build_a_test_stream_pipeline()
-    p.run()
-    self.assertIsNone(
-        ie.current_env().pipeline_result(p, is_main_job=False))
-
-  @patch('apache_beam.runners.interactive.pipeline_instrument.'
-         'PipelineInstrument.has_unbounded_sources', True)
-  def test_background_caching_job_not_start_when_such_job_exists(self):
-    p = _build_a_test_stream_pipeline()
-    a_running_result = runner.PipelineResult(runner.PipelineState.RUNNING)
-    ie.current_env().set_pipeline_result(p, a_running_result, is_main_job=False)
-    main_job_result = p.run()
-    # No background caching job is started so result is still the running one.
-    self.assertIs(a_running_result,
-                  ie.current_env().pipeline_result(p, is_main_job=False))
-    # A new main job is started so result of the main job is set.
-    self.assertIs(main_job_result,
-                  ie.current_env().pipeline_result(p))
-
-  @patch('apache_beam.runners.interactive.pipeline_instrument.'
-         'PipelineInstrument.has_unbounded_sources', True)
-  def test_background_caching_job_not_start_when_such_job_is_done(self):
-    p = _build_a_test_stream_pipeline()
-    a_done_result = runner.PipelineResult(runner.PipelineState.DONE)
-    ie.current_env().set_pipeline_result(p, a_done_result, is_main_job=False)
-    main_job_result = p.run()
-    # No background caching job is started so result is still the running one.
-    self.assertIs(a_done_result,
-                  ie.current_env().pipeline_result(p, is_main_job=False))
-    # A new main job is started so result of the main job is set.
-    self.assertIs(main_job_result,
-                  ie.current_env().pipeline_result(p))
 
 
 if __name__ == '__main__':
